Connection with MySql database Delete automatically order with status DONE
