+++ src/main/java/com/capgemini/molveno/model/Order.java
@@ -2,13 +2,22 @@ package com.capgemini.molveno.model;
 
 import com.capgemini.molveno.enums.OrderStatus;
 
+import javax.persistence.*;
 import java.util.List;
 
-
+@Entity
 public class Order {
-    private List<MenuItem> items;
-    private Table table;
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
     private int Id;
+
+    @ManyToMany(cascade = CascadeType.ALL)
+    private List<MenuItem> items;
+
+    @ManyToOne(cascade=CascadeType.ALL)
+    private Table table;
+
     //it would be ideal if you could set this property for every individual order of the order
     private OrderStatus status;
     //this variable should be composed of the separate prices of every menu-order
